+++ tuk-api/src/main/kotlin/nexters/tuk/application/auth/dto/request/AuthCommand.kt
@@ -1,4 +1,4 @@
-package nexters.tuk.application.auth.dto
+package nexters.tuk.application.auth.dto.request
 
 class AuthCommand {
     sealed class SocialLogin {
