+++ tuk-api/src/main/kotlin/nexters/tuk/application/member/dto/request/MemberCommand.kt
@@ -1,12 +1,11 @@
-package nexters.tuk.application.member.dto
+package nexters.tuk.application.member.dto.request
 
 import nexters.tuk.application.member.SocialType
 
-class MemberResponse {
+class MemberCommand {
     data class SignUp(
-        val memberId: Long,
         val email: String,
-        val socialType: SocialType,
         val socialId: String,
+        val socialType: SocialType,
     )
 }
