+++ tuk-api/src/main/kotlin/nexters/tuk/application/member/dto/response/MemberResponse.kt
@@ -1,11 +1,12 @@
-package nexters.tuk.application.member.dto
+package nexters.tuk.application.member.dto.response
 
 import nexters.tuk.application.member.SocialType
 
-class MemberCommand {
+class MemberResponse {
     data class SignUp(
+        val memberId: Long,
         val email: String,
-        val socialId: String,
         val socialType: SocialType,
+        val socialId: String,
     )
 }
